Move client test to client repo
